+++ ska_particle/src/src/Particle/Impl/TextureGraphicParticleRenderer.cpp
@@ -9,6 +9,6 @@ ska::TextureGraphicParticleRenderer::TextureGraphicParticleRenderer(Renderer& na
 void ska::TextureGraphicParticleRenderer::render(const Renderer&, const ParticleGroup& group) const {
 	const auto& maxSize = group.getLength();
 	for (std::size_t i = 0; i < maxSize; i++) {
-		m_renderer.render(group.appearance, group.pos[i].x, group.pos[i].y, nullptr);
+		m_renderer.render(group.appearance, static_cast<int>(group.pos[i].x), static_cast<int>(group.pos[i].y), nullptr);
 	}
 }
